+++ EDDiscovery/EliteDangerous/JournalEvents/JournalCapShipBond.cs
@@ -18,11 +18,11 @@ namespace EDDiscovery.EliteDangerous.JournalEvents
         {
             AwardingFaction = Tools.GetStringDef(evt["AwardingFaction"]);
             VictimFaction = Tools.GetStringDef(evt["VictimFaction"]);
-            Reward = Tools.GetInt(evt["Reward"]);
+            Reward = Tools.GetLong(evt["Reward"]);
         }
         public string AwardingFaction { get; set; }
         public string VictimFaction { get; set; }
-        public int Reward { get; set; }
+        public long Reward { get; set; }
 
     }
 }
