+++ EDDiscovery/EliteDangerous/JournalEvents/JournalPayLegacyFines.cs
@@ -12,10 +12,10 @@ namespace EDDiscovery.EliteDangerous.JournalEvents
     {
         public JournalPayLegacyFines(JObject evt) : base(evt, JournalTypeEnum.PayLegacyFines)
         {
-            Amount = Tools.GetInt(evt["Amount"]);
+            Amount = Tools.GetLong(evt["Amount"]);
             BrokerPercentage = Tools.GetDouble(evt["BrokerPercentage"]);
         }
-        public int Amount { get; set; }
+        public long Amount { get; set; }
         public double BrokerPercentage { get; set; }
 
         public static System.Drawing.Bitmap Icon { get { return EDDiscovery.Properties.Resources.payfines; } }
